requirements.txt
# Base
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.0
pydantic>=2.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Simulación
numpy>=1.21.0
pandas>=1.3.0
python-json-logger>=2.0.2

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0
pytest-asyncio>=0.15.1

# Frontend deps
websockets>=10.0
aiohttp>=3.8.0

# Utilidades
pyyaml>=5.4.1
python-multipart>=0.0.5

# Development
black>=21.5b2
flake8>=3.9.2
mypy>=0.910

# Added from the code block
aiofiles>=0.7.0 